+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ5KNF3Q1ZSZM1YKWAJCHJDN"
+        "01KJ5V453GFXCPHN0C24EWAN6Q"
     )
     r"""The unique identifier of the data source"""