+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -2193,7 +2193,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KJ5KNEHPPY6BCQWNPFJQ13R9"
+    id: Optional[str] = "01KJ5V44KXPNBX8YMN1BTA4K7C"
 
     description: Optional[str] = None