+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ5KNF3VGFR42V5SAG83XP4H"
+        "01KJ5V453MJW4DE76437Y7MNW5"
     )
     r"""The unique identifier of the data source"""