+++ packages/orq-rc/RELEASES.md
@@ -6029,3 +6029,13 @@ Based on:
 - [python v4.4.0-rc.35] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.35] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.35 - packages/orq-rc
+
+## 2026-02-23 18:13:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.725.0 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.36] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.36] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.36 - packages/orq-rc